big improvements in markets etc

+++ src/Game/Act.java
@@ -6,7 +6,7 @@ import Avatar.SubjectivelyComparable;
 public interface Act extends SubjectivelyComparable{
 	public void doit(Clan doer);
 
-	public int estimateProfit(Clan pOV);
+	public double estimateProfit(Clan pOV);
 
 	public String getDesc();
